+++ realm/realm-library/src/androidTest/java/io/realm/internal/OsResultsTests.java
@@ -47,7 +47,7 @@ import static junit.framework.Assert.fail;
 
 
 @RunWith(AndroidJUnit4.class)
-public class CollectionTests {
+public class OsResultsTests {
     @Rule
     public final TestRealmConfigurationFactory configFactory = new TestRealmConfigurationFactory();
     @Rule
@@ -133,19 +133,19 @@ public class CollectionTests {
     @Test
     public void constructor_withDistinct() {
         SortDescriptor distinctDescriptor = SortDescriptor.getInstanceForDistinct(null, table, "firstName");
-        Collection collection = new Collection(osSharedRealm, table.where(), null, distinctDescriptor);
+        OsResults osResults = new OsResults(osSharedRealm, table.where(), null, distinctDescriptor);
 
-        assertEquals(3, collection.size());
-        assertEquals("John", collection.getUncheckedRow(0).getString(0));
-        assertEquals("Erik", collection.getUncheckedRow(1).getString(0));
-        assertEquals("Henry", collection.getUncheckedRow(2).getString(0));
+        assertEquals(3, osResults.size());
+        assertEquals("John", osResults.getUncheckedRow(0).getString(0));
+        assertEquals("Erik", osResults.getUncheckedRow(1).getString(0));
+        assertEquals("Henry", osResults.getUncheckedRow(2).getString(0));
     }
 
 
     @Test(expected = UnsupportedOperationException.class)
     public void constructor_queryIsValidated() {
-        // Collection's constructor should call TableQuery.validateQuery()
-        new Collection(osSharedRealm, table.where().or());
+        // OsResults's constructor should call TableQuery.validateQuery()
+        new OsResults(osSharedRealm, table.where().or());
     }
 
     @Test
@@ -156,94 +156,94 @@ public class CollectionTests {
         osSharedRealm.commitTransaction();
         // Query should be checked before creating OS Results.
         thrown.expect(IllegalStateException.class);
-        new Collection(osSharedRealm, query);
+        new OsResults(osSharedRealm, query);
     }
 
     @Test
     public void size() {
-        Collection collection = new Collection(osSharedRealm, table.where());
-        assertEquals(4, collection.size());
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
+        assertEquals(4, osResults.size());
     }
 
     @Test
     public void where() {
-        Collection collection = new Collection(osSharedRealm, table.where());
-        Collection collection2 = new Collection(osSharedRealm, collection.where().equalTo(new long[] {0}, oneNullTable, "John"));
-        Collection collection3 = new Collection(osSharedRealm, collection2.where().equalTo(new long[] {1}, oneNullTable, "Anderson"));
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
+        OsResults osResults2 = new OsResults(osSharedRealm, osResults.where().equalTo(new long[] {0}, oneNullTable, "John"));
+        OsResults osResults3 = new OsResults(osSharedRealm, osResults2.where().equalTo(new long[] {1}, oneNullTable, "Anderson"));
 
         // A new native Results should be created.
-        assertTrue(collection.getNativePtr() != collection2.getNativePtr());
-        assertTrue(collection2.getNativePtr() != collection3.getNativePtr());
+        assertTrue(osResults.getNativePtr() != osResults2.getNativePtr());
+        assertTrue(osResults2.getNativePtr() != osResults3.getNativePtr());
 
-        assertEquals(4, collection.size());
-        assertEquals(2, collection2.size());
-        assertEquals(1, collection3.size());
+        assertEquals(4, osResults.size());
+        assertEquals(2, osResults2.size());
+        assertEquals(1, osResults3.size());
     }
 
     @Test
     public void sort() {
-        Collection collection = new Collection(osSharedRealm, table.where().greaterThan(new long[] {2}, oneNullTable, 1));
+        OsResults osResults = new OsResults(osSharedRealm, table.where().greaterThan(new long[] {2}, oneNullTable, 1));
         SortDescriptor sortDescriptor = SortDescriptor.getTestInstance(table, new long[] {2});
 
-        Collection collection2 = collection.sort(sortDescriptor);
+        OsResults osResults2 = osResults.sort(sortDescriptor);
 
         // A new native Results should be created.
-        assertTrue(collection.getNativePtr() != collection2.getNativePtr());
-        assertEquals(2, collection.size());
-        assertEquals(2, collection2.size());
+        assertTrue(osResults.getNativePtr() != osResults2.getNativePtr());
+        assertEquals(2, osResults.size());
+        assertEquals(2, osResults2.size());
 
-        assertEquals(3, collection2.getUncheckedRow(0).getLong(2));
-        assertEquals(4, collection2.getUncheckedRow(1).getLong(2));
+        assertEquals(3, osResults2.getUncheckedRow(0).getLong(2));
+        assertEquals(4, osResults2.getUncheckedRow(1).getLong(2));
     }
 
     @Test
     public void clear() {
         assertEquals(4, table.size());
-        Collection collection = new Collection(osSharedRealm, table.where());
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
         osSharedRealm.beginTransaction();
-        collection.clear();
+        osResults.clear();
         osSharedRealm.commitTransaction();
         assertEquals(0, table.size());
     }
 
     @Test
     public void contains() {
-        Collection collection = new Collection(osSharedRealm, table.where());
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
         UncheckedRow row = table.getUncheckedRow(0);
-        assertTrue(collection.contains(row));
+        assertTrue(osResults.contains(row));
     }
 
     @Test
     public void indexOf() {
         SortDescriptor sortDescriptor = SortDescriptor.getTestInstance(table, new long[] {2});
 
-        Collection collection = new Collection(osSharedRealm, table.where(), sortDescriptor);
+        OsResults osResults = new OsResults(osSharedRealm, table.where(), sortDescriptor);
         UncheckedRow row = table.getUncheckedRow(0);
-        assertEquals(3, collection.indexOf(row));
+        assertEquals(3, osResults.indexOf(row));
     }
 
     @Test
     public void indexOf_long() {
         SortDescriptor sortDescriptor = SortDescriptor.getTestInstance(table, new long[] {2});
 
-        Collection collection = new Collection(osSharedRealm, table.where(), sortDescriptor);
-        assertEquals(3, collection.indexOf(0));
+        OsResults osResults = new OsResults(osSharedRealm, table.where(), sortDescriptor);
+        assertEquals(3, osResults.indexOf(0));
     }
 
     @Test
     public void distinct() {
-        Collection collection = new Collection(osSharedRealm, table.where().lessThan(new long[] {2}, oneNullTable, 4));
+        OsResults osResults = new OsResults(osSharedRealm, table.where().lessThan(new long[] {2}, oneNullTable, 4));
 
         SortDescriptor distinctDescriptor = SortDescriptor.getTestInstance(table, new long[] {2});
-        Collection collection2 = collection.distinct(distinctDescriptor);
+        OsResults osResults2 = osResults.distinct(distinctDescriptor);
 
         // A new native Results should be created.
-        assertTrue(collection.getNativePtr() != collection2.getNativePtr());
-        assertEquals(3, collection.size());
-        assertEquals(2, collection2.size());
+        assertTrue(osResults.getNativePtr() != osResults2.getNativePtr());
+        assertEquals(3, osResults.size());
+        assertEquals(2, osResults2.size());
 
-        assertEquals(3, collection2.getUncheckedRow(0).getLong(2));
-        assertEquals(1, collection2.getUncheckedRow(1).getLong(2));
+        assertEquals(3, osResults2.getUncheckedRow(0).getLong(2));
+        assertEquals(1, osResults2.getUncheckedRow(1).getLong(2));
     }
 
     // 1. Create a results and add listener.
@@ -254,13 +254,13 @@ public class CollectionTests {
         final OsSharedRealm osSharedRealm = getOsSharedRealm();
         Table table = osSharedRealm.getTable("test_table");
 
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        looperThread.keepStrongReference(collection);
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        looperThread.keepStrongReference(osResults);
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection collection1) {
-                assertEquals(collection, collection1);
-                assertEquals(4, collection1.size());
+            public void onChange(OsResults osResults1) {
+                assertEquals(osResults, osResults1);
+                assertEquals(4, osResults1.size());
                 osSharedRealm.close();
                 looperThread.testComplete();
             }
@@ -275,12 +275,12 @@ public class CollectionTests {
         final OsSharedRealm osSharedRealm = getOsSharedRealm();
         Table table = osSharedRealm.getTable("test_table");
 
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection collection1) {
-                assertEquals(collection, collection1);
-                assertEquals(4, collection1.size());
+            public void onChange(OsResults osResults1) {
+                assertEquals(osResults, osResults1);
+                assertEquals(4, osResults1.size());
                 osSharedRealm.close();
                 onChangeCalled.set(true);
             }
@@ -292,17 +292,17 @@ public class CollectionTests {
     @Test
     public void addListener_shouldBeCalledWhenRefreshAfterLocalCommit() {
         final CountDownLatch latch = new CountDownLatch(2);
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        assertEquals(4, collection.size()); // See `populateData()`
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        assertEquals(4, osResults.size()); // See `populateData()`
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection element) {
+            public void onChange(OsResults element) {
                 if (latch.getCount() == 2) {
                     // triggered by beginTransaction
-                    assertEquals(4, collection.size());
+                    assertEquals(4, osResults.size());
                 } else if (latch.getCount() == 1) {
                     // triggered by refresh
-                    assertEquals(5, collection.size());
+                    assertEquals(5, osResults.size());
                 } else {
                     fail();
                 }
@@ -320,11 +320,11 @@ public class CollectionTests {
     @Test
     public void addListener_triggeredByRefresh() {
         final CountDownLatch latch = new CountDownLatch(1);
-        Collection collection = new Collection(osSharedRealm, table.where());
-        collection.size();
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
+        osResults.size();
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection element) {
+            public void onChange(OsResults element) {
                 assertEquals(1, latch.getCount());
                 latch.countDown();
             }
@@ -343,13 +343,13 @@ public class CollectionTests {
         final OsSharedRealm osSharedRealm = getOsSharedRealm();
         Table table = osSharedRealm.getTable("test_table");
 
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        looperThread.keepStrongReference(collection);
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        looperThread.keepStrongReference(osResults);
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection collection1) {
-                assertEquals(collection, collection1);
-                assertEquals(5, collection1.size());
+            public void onChange(OsResults osResults1) {
+                assertEquals(osResults, osResults1);
+                assertEquals(5, osResults1.size());
                 osSharedRealm.close();
                 looperThread.testComplete();
             }
@@ -364,14 +364,14 @@ public class CollectionTests {
         final OsSharedRealm osSharedRealm = getOsSharedRealm();
         Table table = osSharedRealm.getTable("test_table");
 
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        looperThread.keepStrongReference(collection);
-        assertEquals(4, collection.size()); // Trigger the query to run.
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        looperThread.keepStrongReference(osResults);
+        assertEquals(4, osResults.size()); // Trigger the query to run.
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection collection1) {
-                assertEquals(collection, collection1);
-                assertEquals(5, collection1.size());
+            public void onChange(OsResults osResults1) {
+                assertEquals(osResults, osResults1);
+                assertEquals(5, osResults1.size());
                 osSharedRealm.close();
                 looperThread.testComplete();
             }
@@ -389,17 +389,17 @@ public class CollectionTests {
         Table table = osSharedRealm.getTable("test_table");
         final AtomicInteger listenerCounter = new AtomicInteger(0);
 
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        looperThread.keepStrongReference(collection);
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        looperThread.keepStrongReference(osResults);
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection collection1) {
+            public void onChange(OsResults osResults1) {
                 switch (listenerCounter.getAndIncrement()) {
                     case 0:
-                        assertEquals(4, collection1.size());
+                        assertEquals(4, osResults1.size());
                         break;
                     case 1:
-                        assertEquals(5, collection1.size());
+                        assertEquals(5, osResults1.size());
                         osSharedRealm.close();
                         break;
                     default:
@@ -413,9 +413,9 @@ public class CollectionTests {
         looperThread.testComplete();
     }
 
-    private static class TestIterator extends Collection.Iterator<Integer> {
-        TestIterator(Collection collection) {
-            super(collection);
+    private static class TestIterator extends OsResults.Iterator<Integer> {
+        TestIterator(OsResults osResults) {
+            super(osResults);
         }
 
         @Override
@@ -424,8 +424,8 @@ public class CollectionTests {
         }
 
         boolean isDetached(OsSharedRealm osSharedRealm) {
-            for (WeakReference<Collection.Iterator> iteratorRef : osSharedRealm.iterators) {
-                Collection.Iterator iterator = iteratorRef.get();
+            for (WeakReference<OsResults.Iterator> iteratorRef : osSharedRealm.iterators) {
+                OsResults.Iterator iterator = iteratorRef.get();
                 if (iterator == this) {
                     return false;
                 }
@@ -436,8 +436,8 @@ public class CollectionTests {
 
     @Test
     public void collectionIterator_detach_byBeginTransaction() {
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        TestIterator iterator = new TestIterator(collection);
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        TestIterator iterator = new TestIterator(osResults);
         assertFalse(iterator.isDetached(osSharedRealm));
         osSharedRealm.beginTransaction();
         assertTrue(iterator.isDetached(osSharedRealm));
@@ -448,15 +448,15 @@ public class CollectionTests {
     @Test
     public void collectionIterator_detach_createdInTransaction() {
         osSharedRealm.beginTransaction();
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        TestIterator iterator = new TestIterator(collection);
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        TestIterator iterator = new TestIterator(osResults);
         assertTrue(iterator.isDetached(osSharedRealm));
     }
 
     @Test
     public void collectionIterator_invalid_nonLooperThread_byRefresh() {
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        TestIterator iterator = new TestIterator(collection);
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        TestIterator iterator = new TestIterator(osResults);
         assertFalse(iterator.isDetached(osSharedRealm));
         osSharedRealm.refresh();
         thrown.expect(ConcurrentModificationException.class);
@@ -468,13 +468,13 @@ public class CollectionTests {
     public void collectionIterator_invalid_looperThread_byRemoteTransaction() {
         final OsSharedRealm osSharedRealm = getOsSharedRealm();
         Table table = osSharedRealm.getTable("test_table");
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        final TestIterator iterator = new TestIterator(collection);
-        looperThread.keepStrongReference(collection);
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        final TestIterator iterator = new TestIterator(osResults);
+        looperThread.keepStrongReference(osResults);
         assertFalse(iterator.isDetached(osSharedRealm));
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection element) {
+            public void onChange(OsResults element) {
                 try {
                     iterator.checkValid();
                     fail();
@@ -490,29 +490,29 @@ public class CollectionTests {
 
     @Test
     public void collectionIterator_newInstance_throwsWhenSharedRealmIsClosed() {
-        final Collection collection = new Collection(osSharedRealm, table.where());
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
         osSharedRealm.close();
         thrown.expect(IllegalStateException.class);
-        new TestIterator(collection);
+        new TestIterator(osResults);
     }
 
     @Test
     public void getMode() {
-        Collection collection = new Collection(osSharedRealm, table.where());
-        assertTrue(Collection.Mode.QUERY == collection.getMode());
-        collection.firstUncheckedRow(); // Run the query
-        assertTrue(Collection.Mode.TABLEVIEW == collection.getMode());
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
+        assertTrue(OsResults.Mode.QUERY == osResults.getMode());
+        osResults.firstUncheckedRow(); // Run the query
+        assertTrue(OsResults.Mode.TABLEVIEW == osResults.getMode());
     }
 
     @Test
     public void createSnapshot() {
-        Collection collection = new Collection(osSharedRealm, table.where());
-        Collection snapshot = collection.createSnapshot();
-        assertTrue(Collection.Mode.TABLEVIEW == snapshot.getMode());
+        OsResults osResults = new OsResults(osSharedRealm, table.where());
+        OsResults snapshot = osResults.createSnapshot();
+        assertTrue(OsResults.Mode.TABLEVIEW == snapshot.getMode());
         thrown.expect(IllegalStateException.class);
-        snapshot.addListener(snapshot, new RealmChangeListener<Collection>() {
+        snapshot.addListener(snapshot, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection element) {
+            public void onChange(OsResults element) {
             }
         });
     }
@@ -520,15 +520,15 @@ public class CollectionTests {
     @Test
     @RunTestInLooperThread
     public void load() {
-        final Collection collection = new Collection(osSharedRealm, table.where());
-        collection.addListener(collection, new RealmChangeListener<Collection>() {
+        final OsResults osResults = new OsResults(osSharedRealm, table.where());
+        osResults.addListener(osResults, new RealmChangeListener<OsResults>() {
             @Override
-            public void onChange(Collection element) {
-                assertTrue(collection.isLoaded());
+            public void onChange(OsResults element) {
+                assertTrue(osResults.isLoaded());
                 looperThread.testComplete();
             }
         });
-        assertFalse(collection.isLoaded());
-        collection.load();
+        assertFalse(osResults.isLoaded());
+        osResults.load();
     }
 }