+++ realm/realm-library/src/main/java/io/realm/internal/PendingRow.java
@@ -30,7 +30,7 @@ public class PendingRow implements Row {
             "The query has been executed. This 'PendingRow' is not valid anymore.";
 
     private OsSharedRealm osSharedRealm;
-    private Collection pendingCollection;
+    private OsResults pendingOsResults;
     private RealmChangeListener<PendingRow> listener;
     private WeakReference<FrontEnd> frontEndRef;
     private boolean returnCheckedRow;
@@ -38,7 +38,7 @@ public class PendingRow implements Row {
     public PendingRow(OsSharedRealm osSharedRealm, TableQuery query, SortDescriptor sortDescriptor,
                       final boolean returnCheckedRow) {
         this.osSharedRealm = osSharedRealm;
-        pendingCollection = new Collection(osSharedRealm, query, sortDescriptor, null);
+        pendingOsResults = new OsResults(osSharedRealm, query, sortDescriptor, null);
 
         listener = new RealmChangeListener<PendingRow>() {
             @Override
@@ -46,7 +46,7 @@ public class PendingRow implements Row {
                 notifyFrontEnd();
             }
         };
-        pendingCollection.addListener(this, listener);
+        pendingOsResults.addListener(this, listener);
         this.returnCheckedRow = returnCheckedRow;
         osSharedRealm.addPendingRow(this);
     }
@@ -207,8 +207,8 @@ public class PendingRow implements Row {
     }
 
     private void clearPendingCollection() {
-        pendingCollection.removeListener(this, listener);
-        pendingCollection = null;
+        pendingOsResults.removeListener(this, listener);
+        pendingOsResults = null;
         listener = null;
         osSharedRealm.removePendingRow(this);
     }
@@ -224,9 +224,9 @@ public class PendingRow implements Row {
             return;
         }
 
-        if (pendingCollection.isValid()) {
+        if (pendingOsResults.isValid()) {
             // PendingRow will always get the first Row of the query since we only support findFirst.
-            UncheckedRow uncheckedRow = pendingCollection.firstUncheckedRow();
+            UncheckedRow uncheckedRow = pendingOsResults.firstUncheckedRow();
             // Clear the pending collection immediately in case beginTransaction is called in the listener which will
             // execute the query again.
             clearPendingCollection();
@@ -247,7 +247,7 @@ public class PendingRow implements Row {
 
     // Execute the query immediately and call frontend's onQueryFinished().
     public void executeQuery() {
-        if (pendingCollection == null) {
+        if (pendingOsResults == null) {
             throw new IllegalStateException(QUERY_EXECUTED_MESSAGE);
         }